+++ services/data-preparation.service.ts
@@ -1003,10 +1003,7 @@ export const dataPreparationService = {
     await makePolling(
       () =>
         apiClient.purposes.getPurpose(purposeId, getAuthorizationHeader(token)),
-      (res) => {
-        console.log("res:", JSON.stringify(res.data)); // TODO: remove
-        return res.data.waitingForApprovalVersion?.state === "REJECTED";
-      }
+      (res) => res.data.currentVersion?.state === "REJECTED"
     );
   },
 };
